third_party/WebKit/LayoutTests/fast/events/mutation-during-replace-child-2-expected.txt

Ensures that replaceChild() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.replaceChild(newChild, oldChild); threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS successfullyParsed is true

TEST COMPLETE
